Identify Worktree entries by their inode
This will allow us to re-parent elements when re-scanning when the file system changes.

+++ zed/src/editor/buffer/mod.rs
@@ -433,7 +433,7 @@ impl Buffer {
         self.file.as_ref().map(|file| file.path(app))
     }
 
-    pub fn entry_id(&self) -> Option<(usize, usize)> {
+    pub fn entry_id(&self) -> Option<(usize, u64)> {
         self.file.as_ref().map(|file| file.entry_id())
     }
